// src/app/modules/cpassapi/model/impegno.ts
/*
* SPDX-FileCopyrightText: Copyright 2019 - 2020 | CSI Piemonte
* SPDX-License-Identifier: EUPL-1.2
*/
/**
 * Contabilità passiva
 * API per il backend della suite di contabilità passiva.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Ente } from './ente';
import { Fornitore } from './fornitore';
import { Subimpegno } from './subimpegno';


export interface Impegno { 
    id?: string;
    annoEsercizio?: number;
    anno?: number;
    numero?: number;
    descrizione?: string;
    stato?: string;
    importoAttuale?: number;
    importoIniziale?: number;
    liquidatoAnnoPrecedente?: number;
    numeroCapitolo?: number;
    numeroArticolo?: number;
    descrizioneCapitolo?: string;
    annoProvvedimento?: number;
    numeroProvvedimento?: number;
    settoreProvvedimento?: string;
    disponibile?: number;
    pdcCodice?: string;
    pdcDescrizione?: string;
    importo?: number;
    importoRipartito?: number;
    importoSospeso?: number;
    ente?: Ente;
    fornitore?: Fornitore;
    subimpegni?: Array<Subimpegno>;
    optlock?: string;
}