// src/app/modules/cpassapi/model/rigaOrdine.ts
/*
* SPDX-FileCopyrightText: Copyright 2019 - 2020 | CSI Piemonte
* SPDX-License-Identifier: EUPL-1.2
*/
/**
 * Contabilità passiva
 * API per il backend della suite di contabilità passiva.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AliquoteIva } from './aliquoteIva';
import { Destinatario } from './destinatario';
import { Impegno } from './impegno';
import { ListinoFornitore } from './listinoFornitore';
import { Ods } from './ods';
import { StatoElOrdine } from './statoElOrdine';
import { UnitaMisura } from './unitaMisura';


export interface RigaOrdine { 
    id?: string;
    consegnaParziale?: number;
    importoNetto?: number;
    importoIva?: number;
    importoSconto?: number;
    importoSconto2?: number;
    importoTotale?: number;
    importoDaEvadere?: number;
    percentualeSconto?: number;
    percentualeSconto2?: number;
    prezzoUnitario?: number;
    progressivo?: number;
    quantita?: number;
    impegniOrdine?: Array<Impegno>;
    aliquoteIva?: AliquoteIva;
    ods?: Ods;
    statoElOrdine?: StatoElOrdine;
    unitaMisura?: UnitaMisura;
    destinatario?: Destinatario;
    listinoFornitore?: ListinoFornitore;
    optlock?: string;
    note?: string;
}